// FE.Core.FW/src/apis/models/toefl-challenge/ExamLocationModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ExamLocationModel = {
  id?: string;
  examId?: string;
  examScheduleId?: string;
  examScheduleName?: string | null;
  code?: string | null;
  name?: string | null;
  provinceId?: string;
  provinceName?: string | null;
  districtId?: string;
  districtName?: string | null;
  address?: string | null;
  description?: string | null;
  mapLink?: string | null;
};